requirements.txt
# G:\DPR-Net\requirements.txt

torch
torchvision
transformers
open-clip-torch
tqdm
PyYAML
Pillow